Move methods to display text of plans to decorator
