+++ app/controllers/gobierto_plans/plan_types_controller.rb
@@ -24,7 +24,7 @@ module GobiertoPlans
       load_year
       redirect_to gobierto_plans_plan_path(slug: params[:slug], year: @years.first) and return if @year.nil?
 
-      @plan = find_plan
+      @plan = PlanDecorator.new(find_plan)
 
       @site_stats = GobiertoPlans::SiteStats.new site: current_site, plan: @plan
       @plan_updated_at = @site_stats.plan_updated_at
